% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_artists_for_user.R
\name{get_top_artists_for_user}
\alias{get_top_artists_for_user}
\title{get_top_artists_for_user: Get the top artists for the user}
\usage{
get_top_artists_for_user(user_auth_token, limit_num, time_range_opt)
}
\arguments{
\item{user_auth_token:}{a string corresponding to a valid Spotify API user authentication token}

\item{limit_num:}{an integer between 1 and 50 (inclusive)}

\item{time_range_opt:}{an integer, either 1,2 or 3.
1 = `long_term` is last several years of data
2 = `medium_term` is last 6 months
3 = `short_term` last 4 weeks}
}
\value{
a dataframe of artist name, popularity, genres, and followers
     name = character
     popularity = integer between 1 and 100 corresponding to popularity
     genres = character vector of genres associated with artists
     followers = integer of number of followers of artist
}
\description{
get_top_artists_for_user: Get the top artists for the user
}
\examples{
my_auth_token <- "BQDk6Yebn4V94Cbd121Rs73_q4Vtx1WotIX7o3spa5THlzfZi1pQn-jywCkz2XNld5Pa1H3jxo7EGL8JZ-1_G28Sg5as5ZaBeIfKCngrvoT52sLhkUYbplpgk6_G7MIyByENCv32oJAR4txcWgI0j5Bewds-cvR8foos2fz71Xob"
my_df <- get_top_artists_for_user(my_auth_token, 50, 2)
}
