% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related_artists.R
\name{related_artists}
\alias{related_artists}
\title{Prints dataframe of artist's related artists.}
\usage{
related_artists(user_auth_token, artistName)
}
\arguments{
\item{user_auth_token:}{String containing the users authentication tokent. See README for details}

\item{artistName:}{String specifyign an arists name}
}
\value{
dataframe object
}
\description{
Prints dataframe of artist's related artists.
}
\examples{

related_artists(auth, "Haftbefehl")

}
